chore(e2e): add no_demo_data option, property defs, and split page models
- Add `no_demo_data` workspace option to skip demo data generation
- Register property definitions alongside event definitions
- Split insightPage.ts into per-insight files under insights/ folder
