+++ playwright/page-models/insights/pathsInsight.ts
@@ -1,0 +1,13 @@
+import { Locator, Page, expect } from '@playwright/test'
+
+export class PathsInsight {
+    readonly container: Locator
+
+    constructor(page: Page) {
+        this.container = page.getByTestId('paths-viz')
+    }
+
+    async waitForChart(): Promise<void> {
+        await expect(this.container).toBeVisible()
+    }
+}
